requirements.txt
# python -> 3.10

tensorflow-gpu==2.9
numpy==1.26.4
pandas
livelossplot
tqdm
scikit-learn
scipy
ssqueezepy
wfdb
IPython
imbalanced-learn
tensorflow-addons==0.17.0
seaborn==0.13.2
